# server/cmd/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact:
    email: ntwaricliberi@gmail.com
    name: Clarance Liberi Ntwari
    url: https://twitter.com/claranceliberi
  description: This is the API for Ampersand, a car battery swapping service.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Ampersand API
  version: "1.0"
paths:
...
  /swaps/{id}:
    get:
      consumes:
      - application/json
      description: Get Swap
      parameters:
      - description: ID of the swap
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get Swap
      tags:
      - swaps
...
swagger: "2.0"